refactor enums files
