+++ constants/enums.ts
@@ -1,3 +1,14 @@
+export enum STORYBLOK_STORY_STATUS_ENUM {
+  PUBLISHED = 'published',
+  UNPUBLISHED = 'unpublished',
+  DELETED = 'deleted',
+}
+
+export enum LANGUAGES {
+  en = 'en',
+  es = 'es',
+}
+
 export enum PARTNER_ACCESS_CODE_STATUS {
   VALID = 'VALID',
   INVALID_CODE = 'INVALID_CODE',
