--- constants/languages.ts
@@ -1,4 +1,0 @@
-export enum LANGUAGES {
-  en = 'en',
-  es = 'es',
-}
